# docs/swagger.json
{
  "components": {
    "schemas": {
      "data": {
        "title": "data Schema",
        "properties": { "data": { "$ref": "#/components/schemas/Wallet" } },
          "title": "properties Schema",
        "type": "object"
      },
      "ErrorResponse": {
        "title": "ErrorResponse Schema",
        "description": "Application error response",
        "properties": {
          "code": {
            "enum": [400, 401, 404, 500, 502, 422, 403, 409, 429, 501],
            "example": 400,
            "type": "integer"
          },
          "error": { "example": "field: required", "type": "string" },
          "message": {
            "enum": [
              "Invalid request parameters",
              "Authorization failed",
              "Resource not found",
              "Internal server error",
              "Database error occurred",
              "External service error",
              "Error rendering response",
              "Access forbidden",
              "Resource conflict",
              "Too many requests",
              "Unsupported operation"
            ],
            "example": "Invalid request parameters",
            "type": "string"
          },
          "type": { "$ref": "#/components/schemas/ErrorType" }
        },
        "type": "object"
      },
      "ErrorType": {
        "title": "ErrorType Schema",
        "type": "string",
        "x-enum-varnames": [
          "ErrorTypeValidation",
          "ErrorTypeDatabase",
          "ErrorTypeAuthorization",
          "ErrorTypeNotFound",
          "ErrorTypeInternal",
          "ErrorTypeExternalService",
          "ErrorTypeRender",
          "ErrorTypeForbidden",
          "ErrorTypeConflict",
          "ErrorTypeRateLimit",
          "ErrorTypeUnsupported"
        ]
      },
      "Response": {
        "title": "Response Schema",
        "allOf": [{ "$ref": "#/components/schemas/data" }],
        "description": "Standard API response wrapper",
        "properties": {
          "data": {},
          "message": {
            "enum": [
              "Success",
              "Resource created successfully",
              "Resource updated successfully",
              "Resource deleted successfully"
            ],
            "example": "Success",
            "type": "string"
          },
          "meta": {
            "properties": {
              "count": { "type": "integer" },
              "limit": { "type": "integer" },
              "next_token": { "type": "string" },
              "query": { "type": "string" }
            },
            "type": "object"
          },
          "status": {
            "enum": [200, 202, 204],
            "example": 200,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "TagCreatePayload": {
        "title": "TagCreatePayload Schema",
        "description": "Payload for creating a new tag with name and optional color",
        "properties": {
          "color": {
            "example": "#FF5733",
            "format": "hex-color",
            "type": "string"
          },
          "name": {
            "example": "Important",
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": ["name"],
        "type": "object"
      },
      "TagUpdatePayload": {
        "title": "TagUpdatePayload Schema",
        "description": "Payload for updating an existing tag's name and color",
        "properties": {
          "color": {
            "example": "#FF5733",
            "format": "hex-color",
            "type": "string"
          },
          "name": {
            "example": "Important",
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Wallet": {
        "title": "Wallet Schema",
        "description": "A wallet entity",
        "properties": {
          "balance": { "example": 100.5, "type": "number" },
          "createdAt": { "example": "2023-01-01T00:00:00Z", "type": "string" },
          "currency": { "example": "USD", "type": "string" },
          "name": { "example": "My Wallet", "type": "string" },
          "projectId": {
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "type": "string"
          },
          "tags": {
            "items": { "type": "string" },
            "type": "array",
            "uniqueItems": false
          },
          "updatedAt": { "example": "2023-01-01T00:00:00Z", "type": "string" },
          "userId": {
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "type": "string"
          },
          "walletId": {
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Bearer token authentication",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": { "email": "support@example.com", "name": "API Support" },
    "description": "REST API for expense tracking application with user management",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "Expense Tracker API",
    "version": "1.0"
  },
...
  "paths": {
...
    "/tags": {
      "delete": {
        "description": "Deletes all tags for the authenticated user",
        "operationId": "DeleteUserTags",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Response" }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Delete all user tags",
        "tags": ["Tags"]
      },
      "get": {
        "description": "Returns a list of Tags",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/data" }],
                  "description": "Standard API response wrapper",
                  "properties": {
                    "data": {},
                    "message": {
                      "enum": [
                        "Success",
                        "Resource created successfully",
                        "Resource updated successfully",
                        "Resource deleted successfully"
                      ],
                      "example": "Success",
                      "type": "string"
                    },
                    "meta": {
                      "properties": {
                        "count": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "next_token": { "type": "string" },
                        "query": { "type": "string" }
                      },
                      "type": "object"
                    },
                    "status": {
                      "enum": [200, 202, 204],
                      "example": 200,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "List Tags",
        "tags": ["Tags"]
      },
      "post": {
        "description": "Creates a new tag for the authenticated user",
        "operationId": "CreateTag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TagCreatePayload" }
            }
          },
          "description": "Tag creation request",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/data" }],
                  "description": "Standard API response wrapper",
                  "properties": {
                    "data": {},
                    "message": {
                      "enum": [
                        "Success",
                        "Resource created successfully",
                        "Resource updated successfully",
                        "Resource deleted successfully"
                      ],
                      "example": "Success",
                      "type": "string"
                    },
                    "meta": {
                      "properties": {
                        "count": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "next_token": { "type": "string" },
                        "query": { "type": "string" }
                      },
                      "type": "object"
                    },
                    "status": {
                      "enum": [200, 202, 204],
                      "example": 200,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Create a new tag",
        "tags": ["Tags"]
      }
    },
    "/tags/{id}": {
      "delete": {
        "description": "Deletes a tag by ID",
        "operationId": "DeleteTag",
        "parameters": [
          {
            "description": "Tag ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Response" }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Delete a tag",
        "tags": ["Tags"]
      },
      "get": {
        "description": "Retrieves a tag by ID",
        "operationId": "GetTag",
        "parameters": [
          {
            "description": "Tag ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/data" }],
                  "description": "Standard API response wrapper",
                  "properties": {
                    "data": {},
                    "message": {
                      "enum": [
                        "Success",
                        "Resource created successfully",
                        "Resource updated successfully",
                        "Resource deleted successfully"
                      ],
                      "example": "Success",
                      "type": "string"
                    },
                    "meta": {
                      "properties": {
                        "count": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "next_token": { "type": "string" },
                        "query": { "type": "string" }
                      },
                      "type": "object"
                    },
                    "status": {
                      "enum": [200, 202, 204],
                      "example": 200,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Get a tag",
        "tags": ["Tags"]
      },
      "put": {
        "description": "Updates an existing tag",
        "operationId": "UpdateTag",
        "parameters": [
          {
            "description": "Tag ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TagUpdatePayload" }
            }
          },
          "description": "Tag update request",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/data" }],
                  "description": "Standard API response wrapper",
                  "properties": {
                    "data": {},
                    "message": {
                      "enum": [
                        "Success",
                        "Resource created successfully",
                        "Resource updated successfully",
                        "Resource deleted successfully"
                      ],
                      "example": "Success",
                      "type": "string"
                    },
                    "meta": {
                      "properties": {
                        "count": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "next_token": { "type": "string" },
                        "query": { "type": "string" }
                      },
                      "type": "object"
                    },
                    "status": {
                      "enum": [200, 202, 204],
                      "example": 200,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Update a tag",
        "tags": ["Tags"]
      }
    },
...
  },
  "openapi": "3.1.0",
  "servers": [{ "url": "localhost:8080/api/v1" }]
}